refactor: mover vistas y viewModel a la capa de presentación
