+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.DevUApp">
-        <activity android:name=".starwars.StarWarsActivity" />
+        <activity android:name=".starwars.presentation.StarWarsActivity" />
         <activity android:name=".SecondaryActivity" />
         <activity
             android:name=".MainActivity"
