+++ app/src/main/java/com/banregio/devuapp/MainActivity.kt
@@ -14,7 +14,7 @@ import androidx.navigation.ui.setupWithNavController
 import androidx.drawerlayout.widget.DrawerLayout
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
-import com.banregio.devuapp.starwars.StarWarsActivity
+import com.banregio.devuapp.starwars.presentation.StarWarsActivity
 
 class MainActivity : AppCompatActivity() {
