+++ app/src/main/java/com/banregio/devuapp/starwars/presentation/NotificationFragment.kt
@@ -1,4 +1,4 @@
-package com.banregio.devuapp.starwars
+package com.banregio.devuapp.starwars.presentation
 
 import android.os.Bundle
 import android.view.View
