+++ app/src/main/java/com/banregio/devuapp/starwars/presentation/StarWarsViewModel.kt
@@ -1,4 +1,4 @@
-package com.banregio.devuapp.starwars
+package com.banregio.devuapp.starwars.presentation
 
 import android.app.Application
 import android.app.Notification
