+++ app/src/main/res/layout/fragment_films.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".starwars.StarWarsActivity">
+    tools:context=".starwars.presentation.StarWarsActivity">
 
 
     <androidx.constraintlayout.widget.ConstraintLayout
