+++ app/src/main/res/layout/fragment_star_ships.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".starwars.StarShipsFragment">
+    tools:context=".starwars.presentation.StarShipsFragment">
 
 
 </FrameLayout>
