+++ app/src/main/res/layout/fragment_star_wars.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".starwars.StarWarsFragment">
+    tools:context=".starwars.presentation.StarWarsFragment">
 
     <ImageView
         android:id="@+id/iv_header_view"
