+++ app/src/main/res/navigation/star_wars_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/starWarsFragment"
-        android:name="com.banregio.devuapp.starwars.StarWarsFragment"
+        android:name="com.banregio.devuapp.starwars.presentation.StarWarsFragment"
         android:label="@string/lbl_header_star_wars"
         tools:layout="@layout/fragment_star_wars">
         <action
@@ -23,19 +23,19 @@
 
     <fragment
         android:id="@+id/filmsFragment"
-        android:name="com.banregio.devuapp.starwars.FilmsFragment"
+        android:name="com.banregio.devuapp.starwars.presentation.FilmsFragment"
         android:label="FilmsFragment"
         tools:layout="@layout/fragment_films" />
 
     <fragment
         android:id="@+id/starShipsFragment"
-        android:name="com.banregio.devuapp.starwars.StarShipsFragment"
+        android:name="com.banregio.devuapp.starwars.presentation.StarShipsFragment"
         android:label="FilmsFragment"
         tools:layout="@layout/fragment_star_ships" />
 
     <fragment
         android:id="@+id/notificationsFragment"
-        android:name="com.banregio.devuapp.starwars.NotificationFragment"
+        android:name="com.banregio.devuapp.starwars.presentation.NotificationFragment"
         android:label="Notifications"
         tools:layout="@layout/fragment_notification"  />
